requirements.txt
langchain
python-dotenv
Flask
Flask-Cors
faiss-cpu
unstructured
openai
tiktoken